DjangoPrj/requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.5.7
coverage==7.2.5
Django==4.2
django-storages==1.13.2
djangorestframework==3.14.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
outcome==1.2.0
Pillow==9.5.0
psycopg==3.1.8
PySocks==1.7.1
pytz==2023.3
selenium==4.9.1
sniffio==1.3.0
social_auth_app_django==5.2.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==2.0.2
wsproto==1.2.0
factory-boy==3.2.0
python-vkontakte==1.1.2
